+++ packages/react/examples/codesandbox/components/Leadspace/src/index.js
@@ -1,5 +1,5 @@
 /**
- Copyright IBM Corp. 2016, 2021
+ Copyright IBM Corp. 2016, 2022
 
  This source code is licensed under the Apache-2.0 license found in the
  LICENSE file in the root directory of this source tree.
@@ -7,11 +7,9 @@
 
 import './styles.scss';
 
-import React from 'react';
 import ReactDom from 'react-dom';
-
-import { LeadSpace } from '@carbon/ibmdotcom-react';
-import { ArrowRight20 } from '@carbon/icons-react';
+import { LeadSpace } from '@carbon/ibmdotcom-react/es/components/LeadSpace';
+import ArrowRight20 from "@carbon/icons-react/es/arrow--right/20";
 
 const App = () => (
   <div className="bx--grid">
@@ -52,4 +50,4 @@ const App = () => (
   </div>
 );
 
-ReactDom.render(<App/>, document.getElementById('app'));
+ReactDom.render(<App />, document.getElementById('app'));