+++ packages/web-components/src/components/footer/footer-nav.ts
@@ -8,8 +8,8 @@
  */
 
 import { html, customElement, LitElement, property } from 'lit-element';
-import { classMap } from 'lit-html/directives/class-map';
-import settings from 'carbon-components/es/globals/js/settings';
+import { classMap } from 'lit-html/directives/class-map.js';
+import settings from 'carbon-components/es/globals/js/settings.js';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import StableSelectorMixin from '../../globals/mixins/stable-selector';
 import styles from './footer.scss';