+++ packages/web-components/src/components/locale-modal/locale-item.ts
@@ -8,8 +8,8 @@
  */
 
 import { html, property, customElement } from 'lit-element';
-import BXLink from 'carbon-web-components/es/components/link/link';
-import settings from 'carbon-components/es/globals/js/settings';
+import BXLink from 'carbon-web-components/es/components/link/link.js';
+import settings from 'carbon-components/es/globals/js/settings.js';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import styles from './locale-modal.scss';