+++ packages/web-components/src/components/locale-modal/locale-search.ts
@@ -8,7 +8,7 @@
  */
 
 import { html, property, state, query, customElement, LitElement } from 'lit-element';
-import settings from 'carbon-components/es/globals/js/settings';
+import settings from 'carbon-components/es/globals/js/settings.js';
 import { INPUT_SIZE } from 'carbon-web-components/es/components/input/input.js';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import ThrottedInputMixin from '../../globals/mixins/throttled-input';