+++ packages/web-components/src/components/pricing-table/pricing-table-cell-annotation.ts
@@ -8,7 +8,7 @@
  */
 
 import { customElement, LitElement, property, html } from 'lit-element';
-import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings';
+import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
 import StableSelectorMixin from '../../globals/mixins/stable-selector';
 import styles from './pricing-table.scss';