+++ packages/web-components/src/components/structured-list/structured-list-cell.ts
@@ -7,17 +7,22 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import BXStructuredListCell from 'carbon-web-components/es/components/structured-list/structured-list-cell';
+import BXStructuredListCell from 'carbon-web-components/es/components/structured-list/structured-list-cell.js';
 import { customElement, property, html } from 'lit-element';
-import Info16 from 'carbon-web-components/es/icons/information/16';
-import Checkmark20 from 'carbon-web-components/es/icons/checkmark/20';
-import Error20 from 'carbon-web-components/es/icons/error/20';
+import Info16 from 'carbon-web-components/es/icons/information/16.js';
+import Checkmark20 from 'carbon-web-components/es/icons/checkmark/20.js';
+import Error20 from 'carbon-web-components/es/icons/error/20.js';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import DDSStructuredListGroup from './structured-list-group';
 import styles from './structured-list.scss';
 
 const { stablePrefix: ddsPrefix } = ddsSettings;
 
+/**
+ * StructuredListCell
+ *
+ * @element dds-structured-list-cell
+ */
 @customElement(`${ddsPrefix}-structured-list-cell`)
 class DDSStructuredListCell extends BXStructuredListCell {
   parentGroup: DDSStructuredListGroup | null = this.closest(`${ddsPrefix}-structured-list-group`);
@@ -92,4 +97,5 @@ class DDSStructuredListCell extends BXStructuredListCell {
   static styles = styles;
 }
 
+/* @__GENERATE_REACT_CUSTOM_ELEMENT_TYPE__ */
 export default DDSStructuredListCell;