+++ packages/web-components/src/components/structured-list/structured-list-header-row.ts
@@ -7,13 +7,18 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import BXStructuredListHeaderRow from 'carbon-web-components/es/components/structured-list/structured-list-header-row';
+import BXStructuredListHeaderRow from 'carbon-web-components/es/components/structured-list/structured-list-header-row.js';
 import { customElement } from 'lit-element';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import styles from './structured-list.scss';
 
 const { stablePrefix: ddsPrefix } = ddsSettings;
 
+/**
+ * StructuredListHeaderRow
+ *
+ * @element dds-structured-list-header-row
+ */
 @customElement(`${ddsPrefix}-structured-list-header-row`)
 class DDSStructuredListHeaderRow extends BXStructuredListHeaderRow {
   connectedCallback() {
@@ -23,4 +28,5 @@ class DDSStructuredListHeaderRow extends BXStructuredListHeaderRow {
   static styles = styles;
 }
 
+/* @__GENERATE_REACT_CUSTOM_ELEMENT_TYPE__ */
 export default DDSStructuredListHeaderRow;