+++ packages/web-components/src/components/structured-list/structured-list-row.ts
@@ -7,7 +7,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import BXStructuredListRow from 'carbon-web-components/es/components/structured-list/structured-list-row';
+import BXStructuredListRow from 'carbon-web-components/es/components/structured-list/structured-list-row.js';
 import { customElement } from 'lit-element';
 import { html } from 'lit-html';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
@@ -15,6 +15,11 @@ import styles from './structured-list.scss';
 
 const { stablePrefix: ddsPrefix } = ddsSettings;
 
+/**
+ * StructuredListRow
+ *
+ * @element dds-structured-list-row
+ */
 @customElement(`${ddsPrefix}-structured-list-row`)
 class DDSStructuredListRow extends BXStructuredListRow {
   connectedCallback() {
@@ -35,4 +40,5 @@ class DDSStructuredListRow extends BXStructuredListRow {
   static styles = styles;
 }
 
+/* @__GENERATE_REACT_CUSTOM_ELEMENT_TYPE__ */
 export default DDSStructuredListRow;